AP_BattMonitor_Bebop: report raw battery voltage
The battery monitor takes the raw voltage reported by the ESC and applies
filters and a compensation function based on the motor RPMs. This compensated
voltage is also not allowed to ever increase. The compensated voltage is then
used to estimate the battery SOC, which can't be done directly because the
Bebop does not have a current sensor.

Before this commit, only the compensated voltage was reported by the battery
monitor, which made it impossible to see how the battery sags under load. Now,
the raw battery voltage is reported as the primary voltage measurement, and the
compensated voltage is reported as the resting voltage estimate, which
corresponds roughly to what the compensation algorithm attempts to calculate.